AsyncClient refactoring and android-free user tests.
There are now wrappers on Koushi's AsyncHttpClient, AsyncHttpRequest and AsyncHttpResponse, and worker classes taking care of processing the request and the response, so they can be mocked in tests. The Request and Response case classes can be converted to respective AsyncHttp* classes and so they are used directly in AsyncClient.