refactor: tidy up SyncResult usage in sync handlers
This commit removes redundant print statements in all of the failure
cases for the various sync handlers, since the failure logged, tracked
or otherwise reported by the executor of the request.

This commit includes a few stylistic fixes and in some cases fixes
obvious cases where retry should be false.