refactor: increase socket timeouts for long running HttpClient
This client is mainly used for assets, and the default timeout is 10
seconds. This is not very long and can easily be exceeded on bad networks.

Since we would like to fail message sending after 30 seconds, 30 seconds
seems like a good default for now until we can better break up the task
of sending an asset to another client.