refactor: remove interleaving of events in pipeline
Each stage of the pipeline seemed to make the assumption that it would
receive and handle batches of events, however the interleaving stage
actually guaranteed that each stage only ever received one event.

This lead to a steady trickle of notifications, and removing the interleaving
seems to have caused no unwanted issues. For that reason, it's been
removed.