refactor: build notifications from event stream directly
Previously we would wait for changes to the message storage via the
event stream. This, however, meant that the notification service would
often start processesing notifications outside of the pipeline future,
which puts it at risk of being killed by the system if executing in the
background

This commit generates notifications directly from events in the pipeline
so that the generation of notifications will be protected from being killed
by the system, and further more the UI will now be signaled within the
same future instead of publishing to a signal whose subscription may
be executed outside the pipeline.

This helps to separate notifications from the rest of the application
and should also improve testability.