feat: Add storage and support for read receipts
Read receipts are also Confirmation messages like Delivery receipts but with the flag set to READ.
Delivery receipts were also renamed from Receipt to DeliveryReceipt to avoid ambiguity.
A dedicated storage for these receipts was added since read receipts, unlike delivery receipts, will be used in group conversations so a list of users who read a certain message should be stored along with the timestamp of the event.