refactor: remove logged try and static tracking methods
The LoggedTry class encouraged poor dependency usage WRT the tracking
service, and going forward it doesn't give us enough control over what
gets tracked or not.

The tracking service was also changed so as to only ever be injected
where needed so that we improve testability and modularity of the app.

The next step would actually be to move the tracking service impl
to the Android project and merge it with the GlobalTrackingController there.