refactor: turn unused type into wildcard in RefreshingSignal
The value of the refreshing event stream is never used, so we really
don't care what the type is. All we care about is that we have an event
stream of some type, that, when it fires, triggers the load function.

For this reaosn, it's confusing to enforce the use of some type here,
since it indicates it'll be used. This commit just removes that type.