refactor: remove some redundancy in UserSearchServiceSpec
The previous code required the test author to repeat some information in functions call, that could actually be inferred from previous parameters. It was easy to make a mistake and pass in a different information the second time, leading to unexpected test results.

I removed the redundancy, requiring the test author to repeat the following information only once:
- query
- who is the self user

I also removed the need to pass the role for the user. Now it's inferred automatically from the user id.