fix: Ignore markdown links (#548)
There should be no preview displayed for a markdown link.
Unfortunately, we parse the message data in SE and here we prepare links to be displayed with previews, but only much later in UI we apply markdown. When the parsed link part of the message is displayed in UI we no longer know that originally it was a markdown link, and it's difficult to find it out at that point.
Instead, I decided to ignore markdown links during the parsing in SE. It's a bit hacky (now even if there was no markdown handling in UI we would still not display the preview), but it's much easier and faster.