chore: Logs for better investigation of the sync process
All logs starting with 'SYNC' should be treated as temporary. They were added in order to investigate why sometimes the app
gets stuck when syncing on initialization. They can be removed when the bug is fixed.
Apart form adding logs I made two chnages:
1. In `Message`contentUpdater.addContentMessages` I changed `msgs.mapValues` to `msgs.map`.  `mapValues` does not really map
the values - it provides a wrapper with a mapping function, and the mapping takes places only when the value is accessed which
may have side-effects. I'd like to gradually remove `mapValues` form the code.
2. In `CachedStorage` I added a functionality which blocks streams: `onAdded`, `onUpdated`, and `onDeleted`. During sync we
add, update, and delete a lot of data. It might make sense to postpone sending events about it to the rest of the app until
the whole process is done. Right now I use this functionality only when syncing messages. It might be helpful in fixing the
stuck-on-sync bug - iff not, we can remove it later.