refactor: Decreasing the number of future calls in PushService
This is an experimental PR.

We have one device (Meng's) with a dev build connected to the Prod backend. The number of conversations and messages make the sync very slow,
possibly even causing a variant of the stuck-on-sync bug (a race condition due to a lot of parallel syncs?). I decided to simplify `PushService`
the best to my knowledge, limiting the number of created futures and async/awaits. It seems to work a bit faster, but:
1. more through performance testing is needed, both on dev and internal builds
2. the changes in code affect how the full sync is handled; this also needs to be tested.