fix: Self-replies don't show up on Android
When a user from another platform sent a reply to their own message, Android displayed an error message instead.
The bug was regression. It occured after the new assets were merged. In changes to `MessageEventProcessor`
a call to the `checkReplyHashes` method was deleted. The call has to be there, because we need to recreate
locally the hash of the quoted message. If not, the hashes may differ and the quoted message will be treated
as invalid (look to the specs of of the replies feature for details).
The fix simply add back the call to `checkReplyHashes`.