fix: Sync users after the migration to new assets (#568)
Back in April we changed the name of the user preference `ShouldSyncUsers`. The migration code for new assets specified that after the migration the users shoulds be synced, so that their profile pictures will now use new assets. The bug was that the code still used the old name of the preference.