fix: Input and output streams used in image transformation should not point to the same file
While preparing an asset for upload it was possible to get in a corner case where we tried to
transform the image being uploaded (usualy to scale it down) by reading and writing to the
same file in the cache. That resulted in an exception being thrown and the asset not uploaded.