Feature: team logo (#590)
feat: Team logo

This PR includes changes to TeamData required to display the team logo. Specifically, I have changed the type for icon from RAssetId to AssetId to be more aligned with the new way we process assets. Additionally, I've removed the icon_key property since team logos are now considered to be public assets, and therefore the key is no longer needed.

These changes required a migration for both the global and user databases. The migration in the user database is to simply switch a flag so that the team will sync on upgrade. The migration in the global database is to drop the column for icon_key.