fix: Ephemeral assets are not displayed
With new assets the message data was enlarged with a new field, `assetId: Option[AssetId]`.
So, instead of looking through assets for those having the given messageId, we use the fact
that we display at most one asset per message and we can simply store its id in the message.
`MessageEventProcessor` was modified, so that now it looks for assets and sets `assetId` when
creating `MessageData`. However, that part was broken for ephemeral messages: if a message
was ephemeral and having an asset, it was recognized first as having an asset, but then
the code failed to set the asset's id, because the type of the event was invalid.

I suspect the reason for that bug was that `MessageEventProcessor` became very complex. A lot
of changes was made to it during the last years, and it was never refactored. I decided to
rewrite most of it with focus on making it clear how the procedure of turning an event into
a message data looks like.